refactor to map events
